+++ src/onegov.activity/onegov/activity/models/activity.py
@@ -14,7 +14,6 @@ from onegov.user import User
 from sqlalchemy import Column, Enum, Text, ForeignKey
 from sqlalchemy import event
 from sqlalchemy import exists, and_, desc
-from sqlalchemy import Index
 from sqlalchemy.dialects.postgresql import HSTORE
 from sqlalchemy.ext.mutable import MutableDict
 from sqlalchemy.orm import object_session, relationship